Remove PB steps
