--- patchwork/steps/PRPB/PRPB.py
@@ -1,35 +1,0 @@
-from patchwork.step import Step
-from patchwork.steps import PR
-from patchwork.steps.PRPB.typed import PRPBInputs, PRPBOutputs
-
-
-class PRPB(Step, input_class=PRPBInputs, output_class=PRPBOutputs):
-    def __init__(self, inputs):
-        super().__init__(inputs)
-        key_map = dict(path=inputs["path_key"])
-        if inputs.get("comment_title_key") is not None:
-            key_map["commit_message"] = inputs["comment_title_key"]
-        if inputs.get("comment_message_key") is not None:
-            key_map["patch_message"] = inputs["comment_message_key"]
-
-        self.modified_files = []
-        input_modified_files = inputs.get("modified_files")
-        if isinstance(input_modified_files, list):
-            for modified_file in input_modified_files:
-                converted_modified_file = {key: modified_file.get(mapped_key) for key, mapped_key in key_map.items()}
-                if converted_modified_file.get("path") is None:
-                    continue
-                self.modified_files.append(converted_modified_file)
-        elif isinstance(input_modified_files, dict):
-            converted_modified_file = {key: input_modified_files.get(mapped_key) for key, mapped_key in key_map.items()}
-            self.modified_files.append(converted_modified_file)
-        elif isinstance(input_modified_files, str):
-            converted_modified_file = {"path": input_modified_files}
-            self.modified_files.append(converted_modified_file)
-        self.inputs = inputs
-
-    def run(self):
-        pr = PR({**self.inputs, "modified_code_files": self.modified_files})
-        pr_outputs = pr.run()
-        self.set_status(pr.status, pr.status_message)
-        return pr_outputs
