+++ patchwork/steps/SimplifiedLLMOnce/typed.py
@@ -1,17 +1,18 @@
-from typing_extensions import Annotated, Any, Dict, List, TypedDict
+from typing_extensions import Annotated, Any, Dict, TypedDict
 
 from patchwork.common.utils.step_typing import StepTypeConfig
 from patchwork.steps.CallLLM.CallLLM import TOKEN_URL
 
 
-class __SimplifiedLLMOnceInputsRequired(TypedDict):
+class __SimplifiedLLMOncePBInputsRequired(TypedDict):
+    json_schema: Annotated[Dict[str, Any], StepTypeConfig(is_config=True)]
     # PreparePromptInputs
-    prompt_user: Annotated[str, StepTypeConfig(is_config=True)]
+    user_prompt: Annotated[str, StepTypeConfig(is_config=True)]
     prompt_value: Dict[str, Any]
 
 
-class SimplifiedLLMOnceInputs(__SimplifiedLLMOnceInputsRequired, total=False):
-    prompt_system: Annotated[str, StepTypeConfig(is_config=True)]
+class SimplifiedLLMOnceInputs(__SimplifiedLLMOncePBInputsRequired, total=False):
+    system_prompt: Annotated[str, StepTypeConfig(is_config=True)]
     # CallLLMInputs
     model: Annotated[str, StepTypeConfig(is_config=True)]
     openai_api_key: Annotated[
@@ -37,17 +38,3 @@ If you are using a OpenAI API Key, please set `--openai_api_key=<token>`.""",
     google_api_key: Annotated[
         str, StepTypeConfig(is_config=True, or_op=["patched_api_key", "openai_api_key", "anthropic_api_key"])
     ]
-    json: Annotated[bool, StepTypeConfig(is_config=True)]
-    # ExtractModelResponseInputs
-    response_partitions: Annotated[Dict[str, List[str]], StepTypeConfig(is_config=True)]
-
-
-class SimplifiedLLMOnceOutputs(TypedDict):
-    # PreparePromptOutputs
-    prompt: Dict
-    # CallLLMOutputs
-    openai_response: str
-    request_tokens: int
-    response_tokens: int
-    # ExtractModelResponseOutputs
-    extracted_response: Dict[str, str]
