+++ patchwork/steps/__init__.py
@@ -29,26 +29,29 @@ from patchwork.steps.GetTypescriptTypeInfo.GetTypescriptTypeInfo import (
     GetTypescriptTypeInfo,
 )
 from patchwork.steps.JoinList.JoinList import JoinList
-from patchwork.steps.JoinListPB.JoinListPB import JoinListPB
 from patchwork.steps.LLM.LLM import LLM
 from patchwork.steps.ModifyCode.ModifyCode import ModifyCode
-from patchwork.steps.ModifyCodePB.ModifyCodePB import ModifyCodePB
+from patchwork.steps.ModifyCodeOnce.ModifyCodeOnce import ModifyCodeOnce
 from patchwork.steps.PR.PR import PR
 from patchwork.steps.PreparePR.PreparePR import PreparePR
 from patchwork.steps.PreparePrompt.PreparePrompt import PreparePrompt
-from patchwork.steps.PRPB.PRPB import PRPB
 from patchwork.steps.QueryEmbeddings.QueryEmbeddings import QueryEmbeddings
 from patchwork.steps.ReadFile.ReadFile import ReadFile
 from patchwork.steps.ReadIssues.ReadIssues import ReadIssues
 from patchwork.steps.ReadPRDiffs.ReadPRDiffs import ReadPRDiffs
-from patchwork.steps.ReadPRDiffsPB.ReadPRDiffsPB import ReadPRDiffsPB
 from patchwork.steps.ReadPRs.ReadPRs import ReadPRs
 from patchwork.steps.ScanDepscan.ScanDepscan import ScanDepscan
 from patchwork.steps.ScanSemgrep.ScanSemgrep import ScanSemgrep
 from patchwork.steps.SimplifiedLLM.SimplifiedLLM import SimplifiedLLM
 from patchwork.steps.SimplifiedLLMOnce.SimplifiedLLMOnce import SimplifiedLLMOnce
-from patchwork.steps.SimplifiedLLMOncePB.SimplifiedLLMOncePB import SimplifiedLLMOncePB
 from patchwork.steps.SlackMessage.SlackMessage import SlackMessage
+
+# Compatibility Aliases
+JoinListPB = JoinList
+ModifyCodePB = ModifyCodeOnce
+PRPB = PR
+ReadPRDiffsPB = ReadPRDiffs
+SimplifiedLLMOncePB = SimplifiedLLM
 
 __all__ = [
     "AnalyzeImpact",
@@ -73,6 +76,7 @@ __all__ = [
     "LLM",
     "ModifyCode",
     "ModifyCodePB",
+    "ModifyCodeOnce",
     "PR",
     "PreparePR",
     "PreparePrompt",
